+++ backend/src/main/java/de/tomze/backend/service/UserService.java
@@ -12,6 +12,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.EntityNotFoundException;
+import javax.ws.rs.BadRequestException;
 
 import static org.springframework.util.StringUtils.hasText;
 
@@ -117,8 +118,13 @@ public class UserService {
         return userToAppDtoResetPassword;
     }
 
-    public UserEntity deleteUser(String userName) {
+    public UserEntity deleteUser(UserEntity authUser, String userName) {
         UserEntity userEntityToDelete = getUser(userName);
+
+        if(authUser.getRole().equals("admin") && userEntityToDelete.getRole().equals("admin")){
+            throw new IllegalArgumentException("Admin must not delete other admin");
+        }
+
         userRepository.delete(userEntityToDelete);
         return userEntityToDelete;
     }
@@ -130,6 +136,7 @@ public class UserService {
                 .userName(userFromAppDto.getUserName())
                 .password(hashedPassword)
                 .email(userFromAppDto.getEmail())
+                .avatar(userFromAppDto.getAvatar())
                 .build();
     }
 
@@ -139,6 +146,7 @@ public class UserService {
                 .role("user")
                 .userName(userEntity.getUserName())
                 .email(userEntity.getEmail())
+                .avatar(userEntity.getAvatar())
                 .build();
     }
